Adds integration tests for bypass/fall-through response patching behaviors
